Remove blocking from Hash
